Alterhative to 26ea4e: custom Ltl constraint
The other commit solves the same problem: 'interpretLtlAST' *should*
work if all the effects in the 'AST' have an 'InterpretLtl',
'InterpretLtlHigherOrder' or 'InterpretEffectStateful' instance. We can
encode this constraint using a specialized type like the
'ConstraintList' of the "Effect" module.

The drawback is that you'll now explicitly have to type-apply
'interpretLtlAST' to a list of tags signifying which instances you
expect.